services/chat-service/pyproject.toml
[project]
name = "chat-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beanie>=2.0.1",
    "fastapi>=0.133.1",
    "motor>=3.7.1",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "uvicorn>=0.41.0",
    "websockets>=16.0",
]
